+++ src/main/java/org/abg/batchapi/repository/VisitorRepository.java
@@ -1,7 +1,7 @@
-package org.abg.batchapi.repositories;
+package org.abg.batchapi.repository;
 
 
-import org.abg.batchapi.entities.Visitor;
+import org.abg.batchapi.entity.Visitor;
 import org.springframework.data.repository.CrudRepository;
